+++ src/main/java/com/example/gestionoutil/dao/ActeurDAO.java
@@ -30,14 +30,14 @@ public class ActeurDAO {
         return em.createNativeQuery("select * from acteur", MyActeurEntity.class).getResultList();
     }
 
-    public MyActeurEntity getById(int id){
+    public MyActeurEntity getById(Long id){
         EntityManager em = emf.createEntityManager();
         return em.find(MyActeurEntity.class, id);
     }
 
 
 
-    public void delete(int id){
+    public void delete(Long id){
         EntityManager em = emf.createEntityManager();
         EntityTransaction et = em.getTransaction();
